Finished work on the organizer
I believe it works, moves don't overwrite old files. If a name conflict
exists (ie a file was made on the same day with a name previously used)
a number is appended.
Thus if a file "A_picture.jpg" has been moved, and another
"A_picture.jpg" needs to be moved, the second will be called
"A_picture(1).jpg"
This seems pretty standard for this sort of thing. It is now working.

I couldn't figure out a ETAFTP solution for the move so It may have
redundant calls to os.path.exists which is a bummer, but i can't think
of a better way off the top of my head.